Fixed compilation warnings generated by 64-bit MSVC. JIRA: CXX-2411
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@49824 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/connect/services/grid_globals.hpp
@@ -56,7 +56,8 @@ public:
     virtual void Notify(const CWorkerNodeJobContext& job, EEvent event);
 
     void Print(CNcbiOstream& os) const;
-    unsigned int GetJobsRunningNumber() const { return m_ActiveJobs.size(); }
+    unsigned GetJobsRunningNumber() const
+    { return (unsigned) m_ActiveJobs.size(); }
 
     void SetMaxJobsAllowed(unsigned int max_jobs_allowed)
     { m_MaxJobsAllowed = max_jobs_allowed; }
